desktop/requirements.txt
PyQt6>=6.6.0
Pillow>=10.0.0
pyinstaller>=6.0.0
google-genai>=1.0.0
selenium>=4.16.0
webdriver-manager>=4.0.0
